feat: move PhysicalPageFlags to procfs-core
